Add RawLua::create_table_from (internal)
